backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
Pillow>=10.0
reportlab>=4.0
python-decouple>=3.8
drf-yasg>=1.21
django-storages>=1.14
